requirements.txt
requests
sphinx
sphinx_rtd_theme
nose2
transifex-python
click